Updated frontpage
Add buttons to the frontpage whether the user is a Seller or an Admin. If user is an Admin, two buttons were added "List Users" and "Seller Products" allowing admin to monitor users and products (not yet implemented). If user is a Seller, only one button was added: "Add Product" which allows a seller to add their product to the listing (not yet implemented). This helps distinguish different roles amongst our website.